Add whitespace stripping in the input
In HTML sources indentation can aid readability, but it only bloats the
output. This PR adds the argument "strip" to templates, which lets you
automatically strip the template input:

* "none": Don't strip any spaces in the input
* "tail": Remove a single single newline at the end of the input. This
  is the default
* "trim-lines": Remove all whitespaces at the front and back of all
  lines, and remove empty lines
* "eager": Like "trim", but also replace runs of whitespaces with a
  single space.

The automatic stripping does not try to understand the input at all. It
does not know what `xml:space="preserve"`, `<pre>` or `<textarea>`
means.