Reduce askama_derive to a single re-export
All the hard work in askama_derive was actually done in askama_shared.
This PR removes the back-and-forth interaction between the two crates.
Now askama_derive is a single re-export of `#[derive(Template)]` which
has to be done in a proc_macro crate.

This most likely means that askama_derive is "final", unless another
derive template needs to be introduced in the future.