derive: Shrink dependency list
We don't need the more advanced features that `quote` and `syn` offer.
This PR opts out of using them.

Until <https://github.com/rust-lang/rust/issues/92565> is resolved, we
cannot give a better error location than `#[derive(Template)]` itself.
So we don't need to keep track of the span in our `struct CompileError`,
and we can remove the dependency of `proc_macro2`.