Use `try_reserve()` instead of `with_capacity()`
`String::with_capacity()` panics if the requested memory could not be
allocated. `Template::render()` is fallible method, and the fact that it
can panic is not documented.

This commit uses `String::try_reserve()` instead, so even for an
exceedingly large `SIZE_HINT` the method should not panic. In the
generated code `write!()` calls will fail instead with
`Err(std::fmt::Error)`.

I do not test if `try_reserve()` returned an error, because the
rendering might succeed anyway, if less bytes are written than estimated.