book: Add Hyper integration section to Integrations chapter
Adds an explanation of the `with-hyper` feature to the 'Integrations'
chapter. The new section explains the purpose and usage of the `From`
and `TryFrom` implementations provided by the feature, for each
generated template, as well as how the two implementations differ.

It references the integration's example test as the other integration
sections do, but also references the integration's source code for
easier navigation for the reader who wants to dig in a little further.