Make `|tojson` filter faster
This PR changes the filter `|tojson` so that it does not collect the
serialized data into a string, but writes the data into the target
writer directly.

This makes the benchmark run 81% (only serializing) or 39% (serializing
and escaping for HTML) faster. The benchmarked data is not a fair
representation for the data you would most likely serialize, though.