Add versioned book on readthedocs.org
This PR adds the file `.readthedocs.yaml`, which is used by
readthedocs.org to generate the book on their servers. RTD can be used
easily host docs for every Askama version, so users can explore how to
use the current stable version, the current HEAD, and an older version.

For older versions then the current PR, the books won't be built,
because we did not provided the needed `.readthedocs.yaml`.

Also an `index.hbs` is added to made mdbook and RTD work better
together. Only the left sidebar was changed, but you have to override
the whole file to make changes.